Use Twenty Seventeen as parent theme
Switch base styling logic to follow the new Twenty Seventeen theme as a
parent theme. Site title, page title, and widget titles are transformed to
the Flamenco font.

Event icons and files are renamed and removed from actual usage. We're
waiting for the smaller icons to fit in the provided headline space.